+++ src/main/java/proguard/classfile/kotlin/visitor/KotlinValueParameterVisitor.java
@@ -1,7 +1,7 @@
 /*
  * ProGuardCORE -- library to process Java bytecode.
  *
- * Copyright (c) 2002-2020 Guardsquare NV
+ * Copyright (c) 2002-2021 Guardsquare NV
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -51,6 +51,4 @@ public interface KotlinValueParameterVisitor
     {
         visitAnyValueParameter(clazz, kotlinValueParameterMetadata);
     }
-
-    default void onNewFunctionStart() {}
 }